OMAP4 CameraHAL - Compiler warnings removal
This patch removes the compiler warnings thrown during build of CameraHAL components
No functionality impact

+++ camera-omap4/src/FakeCameraAdapter.cpp
@@ -718,7 +718,7 @@ void FakeCameraAdapter::frameThread()
                                         }
                                     else
                                         {
-                                        CAMHAL_LOGEB("Negative refcount for preview buffer 0x%x!", msg.arg1);
+                                        CAMHAL_LOGEB("Negative refcount for preview buffer 0x%x!", (unsigned int)msg.arg1);
                                         }
 
                                 }
@@ -805,7 +805,7 @@ void FakeCameraAdapter::frameThread()
                                 }
                             else
                                 {
-                                CAMHAL_LOGEB("Negative refcount for preview buffer 0x%x!", msg.arg1);
+                                CAMHAL_LOGEB("Negative refcount for preview buffer 0x%x!", (unsigned int)msg.arg1);
                                 }
 
                             }
